Geosyspy - Refacto
Adding MP & VTS services
Adding corresponding unit tests sections
Refacto mrts processor + tests


comment missing
