+++ geosyspy/services/__init__.py
@@ -1,1 +1,3 @@
-__all__ = ["agriquest_service", "analytics_processor_service", "analytics_fabric_service","master_data_management_service","weather_service"]
+__all__ = ["agriquest_service", "analytics_processor_service", "analytics_fabric_service",
+           "master_data_management_service", "weather_service", "vegetation_time_series_service",
+           "map_product_service"]
